[Wasm] Extract binary generation into a separate pipeline phase
KT-77029
