+++ compiler/cli/cli-js/src/org/jetbrains/kotlin/cli/pipeline/web/wasm/WasmOutputGenerationPipelinePhase.kt
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2010-2026 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
+ */
+
+package org.jetbrains.kotlin.cli.pipeline.web.wasm
+
+import org.jetbrains.kotlin.backend.wasm.compileWasmIrToBinary
+import org.jetbrains.kotlin.backend.wasm.linkWasmIr
+import org.jetbrains.kotlin.backend.wasm.writeCompilationResult
+import org.jetbrains.kotlin.cli.pipeline.PerformanceNotifications
+import org.jetbrains.kotlin.cli.pipeline.PipelinePhase
+import org.jetbrains.kotlin.cli.pipeline.web.WasmBackendPipelineArtifact
+import org.jetbrains.kotlin.cli.pipeline.web.WasmIntermediatePipelineArtifact
+import org.jetbrains.kotlin.ir.backend.js.ic.tryAcquireAndRelease
+import org.jetbrains.kotlin.js.config.outputDir
+
+object WasmOutputGenerationPipelinePhase : PipelinePhase<WasmIntermediatePipelineArtifact, WasmBackendPipelineArtifact>(
+    name = WasmOutputGenerationPipelinePhase::class.java.simpleName,
+    preActions = setOf(PerformanceNotifications.BackendStarted),
+    postActions = setOf(PerformanceNotifications.BackendFinished),
+) {
+    override fun executePhase(input: WasmIntermediatePipelineArtifact): WasmBackendPipelineArtifact {
+        (val backendIr, val cacheGuard, val configuration) = input
+        val outputDir = configuration.outputDir!!
+        return cacheGuard.tryAcquireAndRelease {
+            val results = backendIr.map { result ->
+                val linkedModule = linkWasmIr(result)
+                val compileResult = compileWasmIrToBinary(result, linkedModule)
+                writeCompilationResult(
+                    result = compileResult,
+                    dir = outputDir,
+                    fileNameBase = result.baseFileName,
+                    configuration = configuration,
+                )
+                compileResult
+            }
+            WasmBackendPipelineArtifact(results, outputDir, configuration)
+        }
+    }
+}
