+++ compiler/ir/backend.js/src/org/jetbrains/kotlin/ir/backend/js/ic/IncrementalCacheGuard.kt
@@ -1,12 +1,17 @@
 /*
- * Copyright 2010-2023 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Copyright 2010-2026 JetBrains s.r.o. and Kotlin Programming Language contributors.
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
+
+@file:OptIn(ExperimentalContracts::class)
 
 package org.jetbrains.kotlin.ir.backend.js.ic
 
 import org.jetbrains.kotlin.ir.backend.js.ic.IncrementalCacheGuard.AcquireStatus
 import java.io.File
+import kotlin.contracts.ExperimentalContracts
+import kotlin.contracts.InvocationKind
+import kotlin.contracts.contract
 
 class IncrementalCacheGuard(cacheDir: String) {
     enum class AcquireStatus { OK, CACHE_CLEARED, INVALID_CACHE }
@@ -36,15 +41,17 @@ class IncrementalCacheGuard(cacheDir: String) {
 }
 
 inline fun <R> IncrementalCacheGuard.acquireAndRelease(block: (AcquireStatus) -> R): R {
+    contract { callsInPlace(block, InvocationKind.EXACTLY_ONCE) }
     val status = acquire()
     return block(status).also {
         release()
     }
 }
 
-inline fun <R> IncrementalCacheGuard.tryAcquireAndRelease(block: () -> R): R {
-    tryAcquire()
+inline fun <R> IncrementalCacheGuard?.tryAcquireAndRelease(block: () -> R): R {
+    contract { callsInPlace(block, InvocationKind.EXACTLY_ONCE) }
+    this?.tryAcquire()
     return block().also {
-        release()
+        this?.release()
     }
 }
