+++ wasm/wasm.tests/testFixtures/org/jetbrains/kotlin/incremental/WasmAbstractInvalidationTest.kt
@@ -7,7 +7,11 @@ package org.jetbrains.kotlin.incremental
 
 
 import org.jetbrains.kotlin.CoreEnvironmentDeprecation
-import org.jetbrains.kotlin.backend.wasm.*
+import org.jetbrains.kotlin.backend.common.phaser.then
+import org.jetbrains.kotlin.backend.wasm.WasmBackendContext
+import org.jetbrains.kotlin.backend.wasm.WasmCompilerWithICMultimodule
+import org.jetbrains.kotlin.backend.wasm.WasmCompilerWithICSingleModule
+import org.jetbrains.kotlin.backend.wasm.WasmCompilerWithICWholeWorld
 import org.jetbrains.kotlin.backend.wasm.ic.*
 import org.jetbrains.kotlin.backend.wasm.lower.markFunctionToExport
 import org.jetbrains.kotlin.cli.create
@@ -15,6 +19,7 @@ import org.jetbrains.kotlin.cli.jvm.compiler.EnvironmentConfigFiles
 import org.jetbrains.kotlin.cli.jvm.compiler.KotlinCoreEnvironment
 import org.jetbrains.kotlin.cli.pipeline.ConfigurationPipelineArtifact
 import org.jetbrains.kotlin.cli.pipeline.PipelinePhase
+import org.jetbrains.kotlin.cli.pipeline.executePhaseIsolatedWithActions
 import org.jetbrains.kotlin.cli.pipeline.web.WasmIntermediatePipelineArtifact
 import org.jetbrains.kotlin.cli.pipeline.web.WebIncrementalCachePipelineArtifact
 import org.jetbrains.kotlin.cli.pipeline.web.wasm.*
@@ -270,13 +275,8 @@ abstract class WasmAbstractInvalidationTest(
                     verifyCacheUpdateStats(stepId, preparedIcCachesArtifact.dirtyFileLastStats, testInfo + removedModulesInfo)
                 }
 
-                val [parametersList] = incrementalBuildingPhase.executePhase(preparedIcCachesArtifact)!!
-
-                parametersList.forEach { parameters ->
-                    val linkedModule = linkWasmIr(parameters)
-                    val compilationResult = compileWasmIrToBinary(parameters, linkedModule)
-                    writeCompilationResult(compilationResult, buildDir, parameters.baseFileName)
-                }
+                (incrementalBuildingPhase then WasmOutputGenerationPipelinePhase)
+                    .executePhaseIsolatedWithActions(preparedIcCachesArtifact)
             }
 
             when (wasmCompilationMode) {
